Cleanup #3 - Moving from ObjectModel to REST
