--- src/MigrationTools.Clients.AzureDevops.ObjectModel/Processors/PipelineType.cs
@@ -1,10 +1,0 @@
-﻿namespace MigrationTools.Processors
-{
-
-    public enum PipelineType
-    {
-        build,
-        release
-    }
-
-}
