--- src/MigrationTools.Clients.AzureDevops.ObjectModel/Processors/Pipelines/BuildDefinitions.cs
@@ -1,13 +1,0 @@
-﻿using Newtonsoft.Json;
-
-namespace MigrationTools.Processors.Pipelines
-{
-    public partial class BuildDefinitions
-    {
-        [JsonProperty("count")]
-        public long Count { get; set; }
-
-        [JsonProperty("value")]
-        public BuildDefinition[] Value { get; set; }
-    }
-}
