--- src/MigrationTools.Clients.AzureDevops.ObjectModel/Processors/Pipelines/ReleaseBuildDefinitionAbstract.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Newtonsoft.Json;
-
-namespace MigrationTools.Processors.Pipelines
-{
-    public abstract class ReleaseBuildDefinitionAbstract
-    {
-        [JsonProperty("name")]
-        public string Name { get; set; }
-    }
-}
