--- src/MigrationTools.Clients.AzureDevops.ObjectModel/Processors/Pipelines/ReleaseDefinitions.cs
@@ -1,13 +1,0 @@
-﻿using Newtonsoft.Json;
-
-namespace MigrationTools.Processors.Pipelines
-{
-    public partial class ReleaseDefinitions
-    {
-        [JsonProperty("count")]
-        public long Count { get; set; }
-
-        [JsonProperty("value")]
-        public ReleaseDefinition[] Value { get; set; }
-    }
-}
