--- src/MigrationTools.Clients.AzureDevops.ObjectModel/Processors/Pipelines/TaskGroups.cs
@@ -1,13 +1,0 @@
-﻿using Newtonsoft.Json;
-
-namespace MigrationTools.Processors.Pipelines
-{
-    public partial class TaskGroups
-    {
-        [JsonProperty("count")]
-        public long Count { get; set; }
-
-        [JsonProperty("value")]
-        public TaskGroup[] Value { get; set; }
-    }
-}
