+++ src/MigrationTools.Clients.AzureDevops.ObjectModel/ServiceCollectionExtensions.cs
@@ -23,8 +23,6 @@ namespace MigrationTools
             context.AddTransient<TfsTeamSettingsProcessor>();
             //TfsSharedQueries
             context.AddTransient<TfsSharedQueryProcessor>();
-            //TfsPipelines
-            context.AddTransient<TfsPipelineProcessor>();
         }
 
         [Obsolete("This is the v1 Archtiecture, we are movign to V2", false)]
